Simplify stroke-order fallback to Ambulance + dashed axes
Single transparent empty state with a larger Ambulance icon and hover
color. Remove variant gallery and /gallery route.
